Massive overhaul - Represents foundational step towards intuitive maintenance of the class
This commit represents an inappropriately large jump from the previous, but that commit had produced a document that passed format review. It was unclear if this large scale change would be successful so it seemed reasonable to wait to commit until some certainty in its veracity was established.

This is a fundamental overhaul of what had become, out of repeated short term necessity, an accumulation of "hacks" to "make things work" and pass format review. While this commit does not solve every nuanced deficiency, the thesis class is now at a fairly maintainable level. Rather than an unwieldy list of precise changes that were made, these will be summarized here and details left to the physical CHANGELOG.txt file:

+ systematically eliminated many of the "hacks" accumulated over the years, particularly those related to the 'tocloft' package
+ parameterized many hard-coded number/length values and began creating user controls for these
+ applied the latest formatting guidelines to the ToC, LoF, LoT, Appendix (& Subappendix), Bibliography and MAINMATTER page headings and corresponding ToC list items to satisfy the 2023 spacing/indentation, labeling, & numbering requirements.
+ reorganized the file structure, i.e. ordering of control sequences according to groups of related tasks. Control sequences where ordering within the .cls file was crucial in order to not be overwritten were themselves encapsulated within new control sequences that could be expanded at a later point, thereby expanding them there as well as if they had been defined there.
+ Improved on the partitioning scheme within the file,  yielding a more easily navigated .cls file with "section" labels, descriptions, and comments/changes.
+ Heavily expanded upon the internal documentation accompanying the TeX code, including extensive commenting explaining the objectives and/or details of various lines of code.

There remains work to be done towards each of these improvements, but the most important deficiencies in the .cls file have now been resolved. Improvements will continue to be made, but maintenance isn't the daunting challenge requiring a depth of TeX knowledge that it had been up until last week.

The CHANGELOG was updated too, but the TODO list needs further review before being modified, as it may need to be flushed and started all over again.

NOTE: in some cases, the code that has been replaced, particularly code added only this year before the wide overhaul, has simply been commented out rather than removed completely. Previous code is often kept for several commits after it has been replaced (commented out) in case a reversal is need in the future. This code can be reevaluated and removed later. Future commits will address the remnant deficiencies and continue improving the user controls and the overall organizational structure of not just the .cls file, but the overall LaTeX Template.